+++ src/OpenTelemetry.Instrumentation.AspNet/MeterProviderBuilderExtensions.cs
@@ -49,18 +49,6 @@ public static class MeterProviderBuilderExtensions
         configure?.Invoke(options);
 
         builder.AddMeter(AspNetMetrics.InstrumentationName);
-        builder.AddView(instrument =>
-        {
-            if (instrument.Unit == "s" && instrument.Meter.Name == AspNetMetrics.InstrumentationName)
-            {
-                return new ExplicitBucketHistogramConfiguration()
-                {
-                    Boundaries = new double[] { 0, 0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1, 2.5, 5, 7.5, 10 },
-                };
-            }
-
-            return null;
-        });
         return builder.AddInstrumentation(() => new AspNetMetrics(options));
     }
 }